refactor: move Contact helpers into own files
